# swagger-ui-dist/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Descrição dos serviços disponíveis para o módulo de Compras.",
    "version": "1.0.0",
    "title": "Purchase - Microservices"
  },
  "basePath": "/api/purchase",
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/provider": {
      "get": {
        "description": "Requisita todos os Fornecedores",
        "summary": "Requisita todos os Fornecedores",
        "tags": [
          "Provider"
        ],
        "operationId": "Providergetall",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Informações sobre todas as entidades",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/ProviderResponse"
              }
            }
          },
          "400": {
            "description": "Parâmetros inválidos (Erro na descrição)"
          }
        }
      },
      "put": {
        "description": "Altera Fornecedor para estado especificado",
        "summary": "Altera Fornecedor para estado especificado",
        "tags": [
          "Provider"
        ],
        "operationId": "Providerupdate",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ProviderUpdateRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Quantidade de linhas alteradas",
            "schema": {
              "$ref": "#/definitions/UpdateResponse"
            }
          },
          "400": {
            "description": "Parâmetros inválidos (Erro na descrição)"
          }
        }
      },
      "post": {
        "description": "Adiciona novo Fonecedor",
        "summary": "Adiciona novo Fonecedor",
        "tags": [
          "Provider"
        ],
        "operationId": "Providerinsert",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ProviderInsertRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Id da entidade inserida",
            "schema": {
              "$ref": "#/definitions/InsertResponse"
            }
          },
          "400": {
            "description": "Parâmetros inválidos (Erro na descrição)"
          }
        }
      }
    },
    "/provider/{id}": {
      "get": {
        "description": "Requisita o Fornecedor com id especificado",
        "summary": "Requisita o Fornecedor com id especificado",
        "tags": [
          "Provider"
        ],
        "operationId": "Providergetbyid",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Id do Fornecedor usado na filtragem",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "Informações sobre a entidade",
            "schema": {
              "$ref": "#/definitions/ProviderResponse"
            }
          },
          "400": {
            "description": "Parâmetros inválidos (Erro na descrição)"
          }
        }
      },
      "delete": {
        "description": "Deleta Fornecedor com id especificado",
        "summary": "Deleta Fornecedor com id especificado",
        "tags": [
          "Provider"
        ],
        "operationId": "Providerdelete",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Id do Fornecedor usado na filtragem",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "Quantidade de linhas excluídas",
            "schema": {
              "$ref": "#/definitions/UpdateResponse"
            }
          },
          "400": {
            "description": "Parâmetros inválidos (Erro na descrição)"
          }
        }
      }
    },
    "/product": {
      "get": {
        "description": "Requisita todos os Produtos",
        "summary": "Requisita todos os Produtos",
        "tags": [
          "Product"
        ],
        "operationId": "Productgetall",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Informações sobre todas as entidades",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/ProductResponse"
              }
            }
          },
          "400": {
            "description": "Parâmetros inválidos (Erro na descrição)"
          }
        }
      },
      "put": {
        "description": "Altera Produto para estado especificado",
        "summary": "Altera Produto para estado especificado",
        "tags": [
          "Product"
        ],
        "operationId": "Productupdate",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ProductUpdateRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Quantidade de linhas alteradas",
            "schema": {
              "$ref": "#/definitions/UpdateResponse"
            }
          },
          "400": {
            "description": "Parâmetros inválidos (Erro na descrição)"
          }
        }
      },
      "post": {
        "description": "Adiciona novo Produto",
        "summary": "Adiciona novo Produto",
        "tags": [
          "Product"
        ],
        "operationId": "Productinsert",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ProductInsertRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Id da entidade inserida",
            "schema": {
              "$ref": "#/definitions/InsertResponse"
            }
          },
          "400": {
            "description": "Parâmetros inválidos (Erro na descrição)"
          }
        }
      }
    },
    "/product/{id}": {
      "get": {
        "description": "Requisita o Produto com id especificado",
        "summary": "Requisita o Produto com id especificado",
        "tags": [
          "Product"
        ],
        "operationId": "Productgetbyid",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Id do Produto usado na filtragem",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "Informações sobre a entidade",
            "schema": {
              "$ref": "#/definitions/ProductResponse"
            }
          },
          "400": {
            "description": "Parâmetros inválidos (Erro na descrição)"
          }
        }
      },
      "delete": {
        "description": "Deleta Produto com id especificado",
        "summary": "Deleta Produto com id especificado",
        "tags": [
          "Product"
        ],
        "operationId": "Productdelete",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Id do Produto usado na filtragem",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "Quantidade de linhas excluídas",
            "schema": {
              "$ref": "#/definitions/UpdateResponse"
            }
          },
          "400": {
            "description": "Parâmetros inválidos (Erro na descrição)"
          }
        }
      }
    },
    "/provider-product": {
      "post": {
        "description": "Adiciona nova associação entre Fonecedor e Produto",
        "summary": "Adiciona nova associação entre Fonecedor e Produto",
        "tags": [
          "Provider Product"
        ],
        "operationId": "ProviderProductinsert",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ProviderProductInsertRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Id da entidade inserida",
            "schema": {
              "type": "object",
              "parameters": {}
            }
          },
          "400": {
            "description": "Parâmetros inválidos (Erro na descrição)"
          }
        }
      }
    },
    "/provider-product/{providerId}/{productId}": {
      "delete": {
        "description": "Deleta associação entre Fonecedor e Produto com id de fornecedor e produto especificados",
        "summary": "Deleta associação entre Fonecedor e Produto com id de fornecedor e produto especificados",
        "tags": [
          "Provider Product"
        ],
        "operationId": "ProviderProductdelete",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "providerId",
            "in": "path",
            "description": "Id do fornecedor da relação usado na filtragem",
            "required": true,
            "type": "integer"
          },
          {
            "name": "productId",
            "in": "path",
            "description": "Id do produto da relação usado na filtragem",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "Quantidade de linhas excluídas",
            "schema": {
              "$ref": "#/definitions/UpdateResponse"
            }
          },
          "400": {
            "description": "Parâmetros inválidos (Erro na descrição)"
          }
        }
      }
    },
...
  },
  "definitions": {
    "ProviderUpdateRequest": {
      "title": "ProviderUpdateRequest",
      "example": {
        "id": 1,
        "name": "Provider X"
      },
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int32"
        },
        "name": {
          "type": "string"
        }
      },
      "required": [
        "id",
        "name"
      ]
    },
    "ProviderInsertRequest": {
      "title": "ProviderInsertRequest",
      "example": {
        "name": "Provider 1"
      },
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        }
      },
      "required": [
        "name"
      ]
    },
    "ProductUpdateRequest": {
      "title": "ProductUpdateRequest",
      "example": {
        "id": 1,
        "name": "Product X",
        "price": 200.00
      },
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int32"
        },
        "name": {
          "type": "string"
        },
        "price": {
          "type": "number",
          "format": "float"
        }
      },
      "required": [
        "id",
        "name",
        "price"
      ]
    },
    "ProductInsertRequest": {
      "title": "ProductInsertRequest",
      "example": {
        "name": "Product 1",
        "price": 100.00
      },
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "price": {
          "type": "number",
          "format": "float"
        }
      },
      "required": [
        "name",
        "price"
      ]
    },
    "ProviderProductInsertRequest": {
      "title": "ProviderProductInsertRequest",
      "example": {
        "productId": 1,
        "providerId": 1
      },
      "type": "object",
      "properties": {
        "productId": {
          "type": "integer",
          "format": "int32"
        },
        "providerId": {
          "type": "integer",
          "format": "int32"
        }
      },
      "required": [
        "productId",
        "providerId"
      ]
    },
...
    "InsertResponse": {
      "title": "InsertResponse",
      "type": "object",
      "properties": {
        "insertedId": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
    "UpdateResponse": {
      "title": "UpdateResponse",
      "type": "object",
      "properties": {
        "rowsAffected": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
    "ProviderResponse": {
      "title": "ProviderResponse",
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int32"
        },
        "name": {
          "type": "string"
        },
        "isDeleted": {
          "type": "boolean"
        },
        "createdAt": {
          "type": "string",
          "format": "date-time"
        },
        "updatedAt": {
          "type": "string",
          "format": "date-time"
        }
      }
    },
    "ProductResponse": {
      "title": "ProductResponse",
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int32"
        },
        "name": {
          "type": "string"
        },
        "price": {
          "type": "number",
          "format": "float"
        },
        "isDeleted": {
          "type": "boolean"
        },
        "createdAt": {
          "type": "string",
          "format": "date-time"
        },
        "updatedAt": {
          "type": "string",
          "format": "date-time"
        },
        "providers": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ProviderResponse"
          }
        }
      }
    },
...
  },
...
}